refactor: moved debug CLI code to a dedicated module
